// Frontend/cufe-app/src/app/core/models/cufe.model.ts
export interface CufePayload {
  numFac: string;
  fecFac: string;
  horFac: string;
  valFac: string;
  vallmp1: string;
  vallmp2: string;
  vallmp3: string;
  valTot: string;
  nitFE: string;
  numAdq: string;
  ciTec: string;
  tipoAmbiente: string;
  codimp1: string;
  codimp2: string;
  codimp3: string;
}

export interface ResultCufe {
  cufe: string
}